Split mul to separate module
